+++ interop/kotlin-metadata/src/main/kotlin/com/squareup/kotlinpoet/metadata/specs/JvmModifier.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 Square, Inc.
+ * Copyright (C) 2021 Square, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,25 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-pluginManagement {
-  repositories {
-    mavenCentral()
-    gradlePluginPortal()
+package com.squareup.kotlinpoet.metadata.specs
+
+import com.squareup.kotlinpoet.AnnotationSpec
+
+/**
+ * Represents a JVM modifier that is represented as an annotation in Kotlin but as a modifier in
+ * bytecode. Examples include annotations such as [@JvmStatic][JvmStatic] or
+ * [@JvmSynthetic][JvmSynthetic].
+ *
+ * This API is considered read-only and should not be implemented outside of KotlinPoet.
+ */
+public interface JvmModifier {
+  public fun annotationSpec(): AnnotationSpec? {
+    return null
   }
 }
-
-plugins {
-  id("org.gradle.toolchains.foojay-resolver-convention") version("0.8.0")
-}
-
-include(
-  ":kotlinpoet",
-  ":interop:javapoet",
-  ":interop:kotlin-metadata",
-  ":interop:ksp",
-  ":interop:ksp:test-processor",
-)
-
-rootProject.name = "kotlinpoet-root"
-
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")