+++ docs/changelog.md
@@ -3,7 +3,13 @@ Change Log
 
 ## Unreleased
 
-* Fix: Small double and float values were set to 0.0 in %L translation (#1919)
+* **Fix**: Don't expand typealiases of function types to `LambdaTypeName`s in `KSTypeReference.toTypeName()`.
+* **Fix**: Small double and float values were set to 0.0 in %L translation (#1919)
+* **Enhancement**: Make enum entry references in `KSAnnotation.toAnnotationSpec()` and `KSClassDeclaration.toClassName()` more robust.
+* Migrate `kotlinpoet-metadata` to stable `org.jetbrains.kotlin:kotlin-metadata-jvm` artifact for Metadata parsing.
+* Promote `kotlinpoet-metadata` out of preview to stable.
+* Update to Kotlin `2.0.0`.
+* Update to KSP `2.0.0-1.0.22`.
 
 ## Version 1.17.0